refactor: move login logic to separate file
